Add size to super block.

+++ readimage.c
@@ -41,7 +41,7 @@ int main(int argc, char **argv)
     a1fs_superblock *sb = (a1fs_superblock *)(disk);
 
     printf("Super Block:\n");
-    printf("Size: %d\n", sb->size)
+    printf("Size: %d\n", (int)sb->size)
 
         printf("    Address:\n");
     printf("    First Inode Bitmap: %d\n", sb->first_ib);
